estimates-tail/pyproject.toml
[project]
name = "estimates-tail"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
readme = "README.md"
dependencies = [
    "django==5.1.6",
    "daphne==4.1.2",
    "channels[daphne]==4.2.0",
    "psycopg[binary]==3.2.4",
    "django-ninja==1.3.0",
    "django-extensions==3.2.3",
    "ipython==8.32.0",
    "channels_redis==4.2.1"
]

[project.optional-dependencies]
dev = ["ruff==0.9.3", "django-cors-headers==4.7.0"]
